\name{export_concordance_index}
\alias{export_concordance_index}
\title{
Export concordance index
}
\description{
Export the concordance index
}
\arguments{
   \item{samplesA}{...}
   \item{samplesB}{...}
   \item{samplesC}{...}
   \item{samplesD}{...}
   \item{sampleStrata}{...}
   \item{sampleWeights}{...}
   \item{sampleStratumCount}{...}
   \item{outX}{...}
   \item{ratio}{...}
   \item{concordantWeights}{...}
   \item{discordantWeights}{...}
   \item{uninformativeWeights}{...}
   \item{relevantWeights}{...}
}
\value{
No return. Modifies the \code{ratio} argument by reference.
}
\author{
Nicolas De Jay, Simon Papillon-Cavanagh, Benjamin Haibe-Kains
}
\keyword{ methods }
