\name{export_filters_bootstrap}
\alias{export_filters_bootstrap}
\title{
Export filters bootstrap
}
\description{
Export the filters
}
\arguments{
   \item{solutionCount}{...}
   \item{solutionLength}{...}
   \item{dataMatrix}{...}
   \item{priorsMatrix}{...}
   \item{priorsWeight}{...}
   \item{sampleStrata}{...}
   \item{sampleWeights}{...}
   \item{featureTypes}{...}
   \item{sampleCount}{...}
   \item{featureCount}{...}
   \item{sampleStratumCount}{...}
   \item{targetFeatureIndices}{...}
   \item{continuousEstimator}{...}
   \item{outX}{...}
   \item{bootstrapCount}{...}
   \item{miMatrix}{...}
}
\author{
Nicolas De Jay, Simon Papillon-Cavanagh, Benjamin Haibe-Kains
}
\keyword{ methods }
