\name{export_mim}
\alias{export_mim}
\title{
Export mim
}
\description{
Export mim
}
\arguments{
   \item{dataMatrix}{...}
   \item{priorsMatrix}{...}
   \item{priorsWeight}{...}
   \item{sampleStrata}{...}
   \item{sampleWeights}{...}
   \item{featureTypes}{...}
   \item{sampleCount}{...}
   \item{featureCount}{...}
   \item{sampleStratumCount}{...}
   \item{continuousEstimator}{...}
   \item{outX}{...}
   \item{bootstrapCount}{...}
   \item{miMatrix}{...}
}
\value{
Returns the mutual information matrix.
}
\author{
Nicolas De Jay, Simon Papillon-Cavanagh, Benjamin Haibe-Kains
}
\keyword{ methods }
