\name{scores}
\Rdversion{1.1}
\docType{methods}
\alias{scores}
\alias{scores,mRMRe.Filter-method}
\alias{scores,mRMRe.Network-method}
\title{
TODO}
\description{
TODO
}
\usage{
\S4method{scores}{mRMRe.Filter}(object)
\S4method{scores}{mRMRe.Network}(object)
}
\arguments{
  \item{object}{a \code{mRMRe.Filter or mRMRe.Network} object.}
}
\author{
Nicolas De Jay, Simon Papillon-Cavanagh, Benjamin Haibe-Kains
}
\examples{
data(scores)
## maybe str(scores) ; plot(scores) ...
}
\keyword{datasets}
