\name{set_thread_count}
\alias{set_thread_count}
\title{
openMP Thread Count
}
\description{
This methods allows you to set the number of cores currently accessible to openMP
}
\arguments{
   \item{thread_count}{number of OPENMP threads to be used}
}
\author{
Nicolas De Jay, Simon Papillon-Cavanagh, Benjamin Haibe-Kains
}
\keyword{ methods }
