\name{solutions}
\Rdversion{1.1}
\docType{methods}
\alias{solutions}
\alias{solutions,mRMRe.Filter-method}
\alias{solutions,mRMRe.Network-method}
\title{
TODO}
\description{
TODO
}
\usage{
\S4method{solutions}{mRMRe.Filter}(object)
\S4method{solutions}{mRMRe.Network}(object)
}
\arguments{
  \item{object}{a \code{mRMRe.Filter or mRMRe.Network} object.}
}
\author{
Nicolas De Jay, Simon Papillon-Cavanagh, Benjamin Haibe-Kains
}
\examples{
data(solutions)
## maybe str(solutions) ; plot(solutions) ...
}
\keyword{datasets}
