\name{visualize}
\Rdversion{1.1}
\docType{methods}
\alias{visualize}

\title{
TODO}
\description{
TODO
}
\usage{data(visualize)}
\format{
  The format is:
 chr "visualize"
}
\author{
Nicolas De Jay, Simon Papillon-Cavanagh, Benjamin Haibe-Kains
}
\examples{
data(visualize)
## maybe str(visualize) ; plot(visualize) ...
}
\keyword{datasets}
