\name{get.thread.count}
\alias{get.thread.count}
\title{
	openMP Thread Count
}
\description{
This methods allows you to retrieve the number of cores currently accessible to openMP}
\usage{
get.thread.count()
}

\author{
Nicolas De Jay, Simon Papillon-Cavanagh, Benjamin Haibe-Kains
}
\examples{
get.thead.count()
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ parallel }
\keyword{ multithread }% __ONLY ONE__ keyword per line
\keyword{ openmp }