\name{target}
\Rdversion{1.1}
\docType{methods}
\alias{target}
\alias{target,mRMRe.Filter-method}
\alias{target,mRMRe.Network-method}
\title{
TODO
}
\description{
TODO
}
\usage{
\S4method{target}{mRMRe.Filter}(object)
\S4method{target}{mRMRe.Network}(object)
}
\arguments{
  \item{object}{a \code{mRMRe.Filter or mRMRe.Network} object.}
}
\author{
Nicolas De Jay, Simon Papillon-Cavanagh, Benjamin Haibe-Kains
}
\examples{
data(target)
## maybe str(target) ; plot(target) ...
}
\keyword{datasets}
